requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.3.1
colorama==0.4.6
distro==1.9.0
fastapi==0.121.2
frozenlist==1.5.0
h11==0.14.0
httpcore==0.17.3
httptools==0.7.1
httpx==0.24.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
magic-filter==1.0.12
MarkupSafe==3.0.3
multidict==6.2.0
openai==0.28.0
propcache==0.3.1
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.11.2
pydantic_core==2.33.1
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.0
python-multipart==0.0.20
python-telegram-bot==20.3
PyYAML==6.0.3
requests==2.32.3
rsa==4.9.1
sniffio==1.3.1
starlette==0.49.3
Telethon==1.41.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
yarl==1.19.0
